Stop rendering erb for bc apps to build nav
When building navigation menu, we previously rendered the form.yml.erb
files to get the title and description.

This change prevents rendering the erb to build the navigation menu by default.

For "subapps" i.e. desktop configs, you can add a manifest per config with the same
name as the form.yml, but with .manifest.yml as the extension. Otherwise the titleized
form of the filename will be used.

For previous functionality set OOD_RENDER_BATCH_CONNECT_ERB_FOR_NAV=1