avoid rendering erb when generating nav
if Configuration.render_batch_connect_erb_for_nav? is false, we
avoid rendering erb when generating nav by defining form_config as {}

we also define a separate BatchConnect::App#authorized? that doesn't
check the validity of form_config and only ensures if a cluster is defined
the user can submit jobs to it

note: this authorized? method still needs to be fixed to enable specifying
the cluster or clusters used in the manifest since this wouldn't be available
from form_config