Example of altered vnc.html
The (re)connect button is mostly (entirely?) seen by users whose VNC session has ended. As such the ability to reconnect to the session is not actually offered by this control and its utility is limited. This example vnc.html and base.css remove the (re)connect button in favor of a message informing the user that their session has ended, and it is now safe to close the window.

At this time this is not intended for inclusion into OnDemand's master branch.