Set production logging to use stderr
OnDemand does not give the user write permissions to the default Rails logging directory. Rails being magic gracefully warns us of this and falls back to logging to standard error and Nginx picks up the log from there. This commit formalizes that logging to standard error is what we intend to do when in production.